requirements_TF2.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.10.0=py37_0
aiohttp=3.6.3=py37he774522_0
argon2-cffi=20.1.0=py37he774522_1
astor=0.8.1=py37_0
async-timeout=3.0.1=py37_0
async_generator=1.10=py37h28b3542_0
attrs=20.2.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=py37_0
brotlipy=0.7.0=py37he774522_1000
ca-certificates=2020.12.8=haa95532_0
cachetools=4.1.1=py_0
certifi=2020.12.5=py37haa95532_0
cffi=1.14.3=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
colorama=0.4.3=py_0
cryptography=3.1.1=py37h7a1dbc1_0
cudatoolkit=10.1.243=h74a9793_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py37_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py37_0
flake8=3.8.4=py_0
freetype=2.10.3=hd328e21_0
gast=0.2.2=py37_0
google-auth=1.22.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py37he7da953_0
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
imutils=0.5.3=pypi_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.18.1=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
jedi=0.17.2=py37_0
jinja2=2.11.2=py_0
joblib=0.17.0=pypi_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py37_0
jupyterlab_pygments=0.1.2=py_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
keyboard=0.13.5=pypi_0
kiwisolver=1.2.0=py37h74a9793_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.13.0.1=h200bbdf_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=hf4a77e7_3
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.1=py37_0
markupsafe=1.1.1=py37hfa6e2cd_1
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py37hba9282a_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2020.2=256
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
msys2-conda-epoch=20160418=1
multidict=4.7.6=py37he774522_1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37_0
nbformat=5.0.7=py_0
nest-asyncio=1.4.1=py_0
notebook=6.1.4=py37_0
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1i=h2bbff1b_0
opt_einsum=3.1.0=py_0
packaging=20.4=py_0
pandas=1.1.3=py37ha925a31_0
pandoc=2.11=h9490d1a_0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
pickleshare=0.7.5=py37_1001
pillow=7.2.0=py37hcc1f983_0
pip=20.2.3=py37_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0.1=py37ha925a31_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.6.0=py_0
pycparser=2.20=py_2
pyflakes=2.2.0=py_0
pygments=2.7.1=py_0
pyjwt=1.7.1=py37_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pyrsistent=0.17.3=py37he774522_0
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
pytz=2020.1=py_0
pywin32=227=py37he774522_1
pywinpty=0.5.7=py37_0
pyzmq=19.0.2=py37ha925a31_1
qt=5.9.7=vc14h73c81de_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.2=py37h9439919_0
send2trash=1.5.0=py37_0
setuptools=50.3.0=py37h9490d1a_1
sip=4.19.8=py37h6538335_0
six=1.15.0=py_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.2.1=pyh532a8cf_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.1.0=gpu_py37h7db9008_0
tensorflow-base=2.1.0=gpu_py37h55f5790_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
termcolor=1.1.0=py37_1
terminado=0.9.1=py37_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=he774522_0
tornado=6.0.4=py37he774522_1
traitlets=5.0.4=py_0
urllib3=1.25.10=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=0.16.1=py_0
wheel=0.35.1=py_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.12.1=py37he774522_1
xz=5.2.5=h62dcd97_0
yarl=1.6.2=py37he774522_0
zeromq=4.3.2=ha925a31_3
zipp=3.3.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
